style: Move helper test files
