+++ packages/runtime-core/__tests__/helpers/renderList.spec.ts
@@ -1,4 +1,4 @@
-import { renderList } from '../src/helpers/renderList'
+import { renderList } from '../../src/helpers/renderList'
 
 describe('renderList', () => {
   it('should render nodes in an array', () => {
